+++ NEWS.md
@@ -9,6 +9,9 @@
   `plot.title.size` or `axis.text.size` now scale with the `base_size` argument,
   to ensure proper scaling of plots when the base-size is changed.
 
+* Added a new theme, `theme_azurelight()`, with a light-blue character, and
+  reduced use of grid lines.
+
 * `theme_modern()` now recognizes a `show.ticks` argument, to add tick marks
   to the plot and slightly increase the distance between axis labels and the
   related axis.